move ParametersSection to enterprise tree
