+++ enterprise/frontend/src/metabase-enterprise/sharing/index.js
@@ -1,0 +1,4 @@
+import { PLUGIN_DASHBOARD_SUBSCRIPTION_PARAMETERS_SECTION_OVERRIDE } from "metabase/plugins";
+import MutableParametersSection from "metabase-enterprise/sharing/components/MutableParametersSection";
+
+PLUGIN_DASHBOARD_SUBSCRIPTION_PARAMETERS_SECTION_OVERRIDE.Component = MutableParametersSection;
