+++ frontend/src/metabase/plugins/index.js
@@ -49,3 +49,7 @@ export const PLUGIN_SNIPPET_SIDEBAR_PLUS_MENU_OPTIONS = [];
 export const PLUGIN_SNIPPET_SIDEBAR_ROW_RENDERERS = {};
 export const PLUGIN_SNIPPET_SIDEBAR_MODALS = [];
 export const PLUGIN_SNIPPET_SIDEBAR_HEADER_BUTTONS = [];
+
+export const PLUGIN_DASHBOARD_SUBSCRIPTION_PARAMETERS_SECTION_OVERRIDE = {
+  Component: undefined,
+};
